User Story 9 Complete
Players can now finish a game when one player's HP reaches 0.
Updated Game to include #game_over? to check whether attacked players'
HP is 0.  Attack is now sent to a POST/attack route which calls the
 #game_over? method.  IF game is complete, the user is redirected to
GET'/game-over' and a closing message is displayed.  Otherwise, the User
is directed to GET/attack and the game continues.